crud completed

+++ src/App/Controller/BookController.php
@@ -4,11 +4,9 @@ namespace App\Controller;
 
 use App\Entity\Book;
 use App\Entity\Author;
-use App\Entity\Manufacturer;
 use App\Exception\ApiProblemException;
-use Bezhanov\Silex\Routing\Route;
+//use Bezhanov\Silex\Routing\Route;
 use Hateoas\Representation\Factory\PagerfantaFactory;
-use Monolog\Logger;
 use Pagerfanta\Adapter\DoctrineORMAdapter;
 use Pagerfanta\Pagerfanta;
 use Symfony\Component\HttpFoundation\Request;
@@ -19,16 +17,26 @@ class BookController extends ResourceController
     /**
      * @Route("/books", methods={"GET"}, name="list_books")
      */
-    public function indexAction(Request $request)  //perhaps delete pagination
+    public function indexAction(Request $request) //?title={title}
     {
-        $queryBuilder = $this->em->createQueryBuilder()->select('f')->from($this->getEntityClassName(), 'f')->addOrderBy('f.id');
-        $adapter = new DoctrineORMAdapter($queryBuilder);
-        $pager = new Pagerfanta($adapter);
-        $pager->setCurrentPage($request->query->get('page', 1))->setMaxPerPage($request->query->get('limit', 30));
-        $factory = new PagerfantaFactory();
-        $collection = $factory->createRepresentation($pager, new \Hateoas\Configuration\Route('list_books'));
+        $title = $request->get('title');
+        if($title) {
+            $queryBuilder = $this->em->createQueryBuilder()
+                ->select('b')
+                ->from($this->getEntityClassName(), 'b')
+                ->where('b.title like :title')
+                ->setParameter('title', '%' . $title . '%')
+                ->addOrderBy('b.id');
+        } else {
+            $queryBuilder = $this->em->createQueryBuilder()->select('b')->from($this->getEntityClassName(), 'b')->addOrderBy('b.id');
+        }
+            $adapter = new DoctrineORMAdapter($queryBuilder);
+            $pager = new Pagerfanta($adapter);
+            $pager->setCurrentPage($request->query->get('page', 1))->setMaxPerPage($request->query->get('limit', 30));
+            $factory = new PagerfantaFactory();
+            $collection = $factory->createRepresentation($pager, new \Hateoas\Configuration\Route('list_authors'));
 
-        return $this->createApiResponse($collection, Response::HTTP_OK);
+            return $this->createApiResponse($collection, Response::HTTP_OK);
     }
 
     /**
@@ -37,7 +45,7 @@ class BookController extends ResourceController
     public function createAction(Request $request)
     {
         $expectedParameters = ['title', 'description'];
-        $requestBody =$request->request->all();
+        $requestBody = $request->request->all();
 
         $book = new Book();
 
@@ -48,14 +56,12 @@ class BookController extends ResourceController
             throw new ApiProblemException(ApiProblemException::TYPE_VALIDATION_ERROR);
         }
 
-        if($requestBody['author']){
-            foreach($requestBody['author'] as $k=>$v ){
-                $author = new Author();
-                $author->setName($requestBody['author'][$k]['name']);
-                $author->setSurname($requestBody['author'][$k]['surname']);
-                $book->addAuthor($author);
-                $this->em->persist($author);
-                //return print_r($book);
+        if ($requestBody['author']) {
+            foreach ($requestBody['author'] as $k => $v) {
+                $author = $this->em->find(Author::class, $requestBody['author'][$k]['id']);
+                if($author){
+                    $book->addAuthor($author);
+                }
             }
         }
 
@@ -86,12 +92,7 @@ class BookController extends ResourceController
         $expectedParameters = ['title', 'description'];
         $requestBody = $request->request->all();
 
-        //$requestBody = $this->extractRequestBody($request, $expectedParameters);
-        //$requestBody['manufacturer'] = $this->em->getReference(Manufacturer::class, $requestBody['manufacturer_id']);
-        //array_splice($expectedParameters, -1, 1, ['manufacturer']);
-
         $book = $this->findOrFail($id);
-
         $book->fromArray($requestBody, $expectedParameters);
         $violations = $this->validator->validate($book);
 
@@ -99,20 +100,10 @@ class BookController extends ResourceController
             throw new ApiProblemException(ApiProblemException::TYPE_VALIDATION_ERROR);
         }
 
-        if($requestBody['author']){
-            foreach($requestBody['author'] as $k=>$v ){
-               /*$author = //$em->find('Author', $requestBody['author'][$k]['id']);//get from authorContr
-               //$author = new Author(4);
-                $name = $requestBody['author'][$k]['name'] ?? null;
-                $surname = $requestBody['author'][$k]['surname'] ?? null;
-                if ($name){
-                    $author->setName($name);
-                }
-                if ($surname){
-                    $author->setSurname($surname);
-                }
-                //$book->addAuthor($author);
-                $this->em->persist($author);*/
+        foreach ($requestBody['author'] as $k => $v) {
+            $author = $this->em->find(Author::class, $requestBody['author'][$k]['id']);
+            if (!$book->getAuthors()->contains($author)) {
+                $book->addAuthor($author);
             }
         }
 
@@ -127,6 +118,11 @@ class BookController extends ResourceController
     public function deleteAction(int $id)
     {
         $book = $this->findOrFail($id);
+
+        if ($book->getOrders()->count()) {
+            return $this->createApiResponse('You should remove connected orders first', Response::HTTP_OK);
+        }
+
         $this->em->remove($book);
         $this->em->flush();
